WIP - Dockerizing development environment
This is mainly to avoid using RVM and dependencies. Not sure whether this affects the build on Github's end of things.

I have already played around with changing Ruby versions here, and while currently the version of Jekyll that Github Pages uses still supports Ruby >=2.5, we are probably not far away from requiring an upgrade, so a good idea to do it now.